Who even needs CSS?
Minimize the amount of CSS even more since the site was almost starting
to look designed. I actually started this evening planning to rebuild
the site and use fancy fonts and things, and this is where we are now so
that should be an indication of how well that went.