Add basic albums, databag, photoset support
This adds photo albums to the site, in a crude form. Originally I built
this using Postgres, and added a whole database layer to the app,
including a hand-rolled migration script.... But then I realised that
was silly, so redid it from scratch to use YAML.

The DataBag implementation isn't inherently specific to Photos/Albums,
so I'd like to try and make articles conform to this data API as well.