Add DataBag for pages; move DPI app into it
Just noticed that the "pages controller" was exclusively for serving 1
specific page (lol). This updates it to pull from a new DataBag for
"pages" so it can be generically used. To prove that point, we relocate
the "DPI calculator" page out of its own bespoke controller and under
the new Pages one.