Add native support for ToC
This adds proper Table of Contents support via YML metadata. It removes
the normal container and makes the article a bit wider to accomodate it,
turning the whole thing into a two-column layout. We also get to cheat
and use `position: sticky` now which made my wildest dreams come true.