Import some Hanami bits and use it by default
This imports some parts of Hanami for routing and views, and then sets
that up as the Rack app that starts, but set it to proxy all requests to
the legacy Yokunai application.

This allows us to do a gradual migration of legacy website bits to the
new stack without having to do a giant code dump.