Replace root controller with Hanami
The first route moved from yokunai to hanami :tada:

This sets up a few things in addition to the root controller, as the
Hanami libraries were not configured properly previously.

We did have to duplicate the layout, as some syntax and locals-access
changed, but that should be fine as a temporary measure.