Add initial site
This accidentally is the first commit where I just build the whole site
in one go.

This is my first attempt using Gatsby for anything. I imported a large
portion of content from the existing site, though tweaked and omitted
some of it.

Still a ways to go -- I haven't even loaded the site on a mobile device
so I have no idea if it even works there. Some content styles are a bit
rough still as well.