Add production image and fix scripts for GNU
Until now we have only been running this on macOS, which has a BSD
userland, so some tweaks were needed ffor `cp` and `date`, which have
different behaviours across the two implementations.

For production, this will be running on Kubernetes, so we need to
provide a Docker image, which just serves the static _build files.