Switch to a makefile
This replaces a lot of the original script with just a Makefile.
Ultimately, recreating what Make does in a script was just not worth it.
We should just use make.

This also reorganizes the code a lot by splitting it into various
"categories" -- one folder for each "type" of content. I think this
works a lot better.