Read files in restrictive mode
This limits a bunch of dangerous or unwanted things while parsing the
files themselves. This makes using bash as a template language slightly
safer, though there is still a large chance for something to execute
that you didn't want.

The most annoying part of restrictive mode is you can't redirect output,
which we were using to silence `date --version`. We're effectively doing
the same thing now just by grepping explicitly for a GNU string, instead
of trusting the return value of a command directly.