Update setup page with XPS 13
I sold my Macbook, after having it need a full "top case" replacement
because of a faulty keyboard, because I simply lost faith that it would
last any appreciable amount of time without breaking yet again.

The XPS 13 is some solid competition, with an equally (if not more)
beautiful HiDPI screen, a keyboard with actual key travel, an even
smaller footprint, thinner bezels... and as a bonus I can run Linux.

it's good™