Refactor apt complains article
As is almost universally true, this post didn't sit well with me a
couple days after writing it. I wrote it fresh out of the frustration
caused by the issue I ran into, and it became a rant piece that did no
one any favours.

This is my attempt to save it, by being a bit less aggressive, a bit
less ranty, and more positive overall -- the point in my head originally
was to extol the wonders of dnf, but that was essentially a footnote.
It still is more or less a footnote, but at least you might actually get
there before writing me off as an arrogant crank. Maybe.