Use Drone promotions instead of deploy branch
This is a more "explicit" way to do the same thing -- we can use the
Drone CLI to "promote" a build, which rebuilds it and runs the steps
which correspond to that promotion. In this case, we can "promote" a
build to production and it'll rebuild and run the rollout steps for that
promoted build.