Dark mode support
This moves all colours into variables so that we can swap them out in
the "pls dark" media query, changing the entire site's colour scheme
with minimal effort and duplication.

I tried my best but because I'm not a dark mode person I don't know if I
bungled the standards or typical patterns for this sort of UI design.
But at least it won't blind people anymore.

This also updates the code colour schemes: Dracula for dark mode, and
trac for light mode. Makes both a little more lively.