Set Cache-Control on uploaded blobs
This will help the CDN cache things more effectively, as currently we're
getting significantly more origin hits than cache hits.

Setting the max-age to 2 days, as that is long enough to prevent most
origin hits while also letting me be lazy and not to manual
invalidations when I want to deploy new things.