New year, new setup
I switched back to macOS recently with the intention of buying back into
the Apple ecosystem in time to eventually upgrade to an ARM Macbook.

There were also a bunch of old things or missing pieces of the setup
page, so this overhauls a bunch of it unrelated to the Macbook parts.