Slim down and simplify development container
We should clean up some of the development Docker container stuff to
not rely on scripting languages, and pack as much as possible into the
container itself.

1. Remove Ruby and Python: replace the http server with nginx, pygments
   with chroma, and foreman with shoreman.
2. Remove the bin/dev script by putting the templates it generates into
   the Docker container itself and spawn shoreman directly as the CMD.
3. Upgrade alpine to the latest release for good luck.

The resulting container size savings are rather significant: from ~70MB
down to just ~20MB.

This will also mean that we will need to switch to chroma for
non-container builds as well, but this shouldn't be too big of a hassle,
and one less reason to use python is always good with me.