Remove Yosemite article
This was from a long time ago, I wrote it while I was young and angry.
It's a pretty disgusting article in tone and content. Just a bunch of
shitting on things, swearing, and being a dick.

This shouldn't have been published, and certainly has no place to be
published anymore, especially attributed to my name.