Add bin/dev for using container for development
This automates the annoying process of typing a long docker-run string
every time I want to just run a dev server, only to realise I
accidentally cleaned out the custom image and have to rebuild it from
the secondary Dockerfile... `bin/dev` is much easier.

Also updated the README to mention chroma instead of pygments, and
updated the ag|entr example to just use find as we make a point of only
requiring unix tools, so why add an unnecessary one, especially since
find is used by the Docker container.