Reintroduce Setup page
I removed this during the last redesign because it was out of date and I
didn't want to redo the whole thing, plus I was going for whole-site
minimalism so the fewer pages the better.

I've had to field a couple questions recently that reminded me why I had
the page to begin with. So it's back, with a slightly different layout,
and fully up-to-date information.

Since this is the first non-article page in the new redesign, I had to
create a global nav menu, which required a few tweaks to make it work on
the homepage.