Add disclaimer note to nextcloud article
Nextcloud is not cloud-native and forcing it into that paradigm is a lot
of hassle, risk, and annoyance. I am much happier to just have it
running on bare metal, with all my files in a normal on-disk format, and
updates just a click away without worrying if i'm going to corrupt the
whole installation when the pod doing the update gets evicted in the
middle of it.

I hope one day I can run something that doesn't behave like it's 1999,
but until then, let's get physical.