Add optional webp image source on homepage
The portrait is the heaviest part of the homepage (being the only image)
at around 50kb. This can be reduced a fair bit by offering a webp
version for modern browsers. This reduces the weight to just 30kb.

This also sets explicit width/height on the fallback img for good luck.

And since the picture element is styled slightly differently, we set
line-height to zero to remove phantom whitespace that showed up beneath
it, that the img did not have because of the flex properties.