Improve cross-platform compatibility
This breaks out the chroma install into a script so we can do a bit of
`uname` sleuthing to download different OS and architecture versions of
chroma instead of hardcoding the linux x64 version.

While we're trying to make things compatible, let's finally bring back
the BSD coreutils branching where we're parsing dates. The GNU/BSD date
differences are the only thing that is preventing builds on macOS
currently, so it seems worth it to just keep the tiny little branches in
the couple places they're needed.

Since this fixes building the app on macOS and also on arm, let's just
kill the overcomplicated Docker dev environment.