Generate canonical permalinks automatically
Recently I fixed a bad redirect bug caused by the CDN not sending the
right headers, which meant certain nginx redirects would rewrite the
hostname to be the origin subdomain, which shouldn't ever be public.

This reminded me that I really need to set up canonical tags to ensure
search engines don't accidentally end up on the CDN origin or something
similar. But I'm not going to write them by hand, so we need to just
generate them based on the filename.

We were already hard-coding permalinks in a bunch of places for
opengraph metadata, so this also will clean up all of those which is
nice. The base URL is still hardcoded, but now only in one place, which
makes it much easier to fix later. I have left a TODO comment to that
effect.