Convert site to custom Go application
This is a big spike of work to replace the somewhat-janky Makefile and
shell script workflow with a Go web server and content parsing pipeline.
This will allow for more formats of content -- namely markdown -- as
well as give a bit more flexibility in how things are built, and how
rich our features can get.

Before anyone asks, yes, this was the amount of work I chose over just
using Hugo, even thought Hugo was probably the more correct choice.