Add Repository <-> Rendering abstraction
This moves a lot of stuff around and introduces a new abstraction: the
repository. This pushes all the filesystem access stuff into a single
place with a standardised interface so that -- theoretically -- we could
replace the filesystem storage with a database or even some other thing
without having to change the entire rendering pipeline.

By doing this, we separate the _rendering_ of the content from the
_storage_ of it, which feels much cleaner and less like a big pile of
spaghetti.