Combine code stylesheets into one
It's not that difficult to just wrap the dark mode colour scheme in an
`@media` tag instead of an entirely separate stylesheet. I think
originally I did this because I wanted to generate them as output from
some other tool like pygmentize, but those days are long gone and these
are pretty untouched, so just maintaining both in one file manually is not
a burden.