Add site-wide Config struct
We need to be able to determine if the program is running in production
or not, as well as store some global configuration such as a paths,
branding, and base URLs.

We should introduce a Config struct to hold this site-wide configuration
and pass it through wherever it's needed.

This will fix analytics in production as we will now have a way to set
and check the app profile, and will fix some hard-coded URLs in
templates.