Add a proper 'not found' page
Currently if something is not found we just dump a 500 and json error in
the response which is pretty gross and Google is complaining about it
because some ancient URLs are now returning 500.

We should add a proper '404 not found' page template, and render it when
we fail to find a piece of content.