Add old post alert
Imported a few ancient blog posts but I never added this little warning
back to the site in whatever redesign this was many years ago.

Just feels nice to have a little disclaimer since some of my ancient
posts still get a fair bit of traffic and for some of them I was a
literal child, so... a grain-of-salt warning seems in order.